Move beer_lambert and optical_density to nirs submodule
