+++ mne/io/nirx/tests/test_nirx.py
@@ -13,8 +13,8 @@ from mne.io import read_raw_nirx
 from mne.io.tests.test_raw import _test_raw_reader
 from mne.transforms import apply_trans, _get_trans
 from mne.utils import run_tests_if_main
-from mne.preprocessing.nirs import nirs_source_detector_distances,\
-    nirs_short_channels
+from mne.preprocessing.nirs import source_detector_distances,\
+    short_channels
 
 fname_nirx_15_0 = op.join(data_path(download=False),
                           'NIRx', 'nirx_15_0_recording')
@@ -47,7 +47,7 @@ def test_nirx_15_2_short():
     # nirsite https://github.com/mne-tools/mne-testing-data/pull/51
     # step 4 figure 2
     allowed_distance_error = 0.0002
-    distances = nirs_source_detector_distances(raw.info)
+    distances = source_detector_distances(raw.info)
     assert_allclose(distances[::2], [
         0.0304, 0.0078, 0.0310, 0.0086, 0.0416,
         0.0072, 0.0389, 0.0075, 0.0558, 0.0562,
@@ -56,11 +56,11 @@ def test_nirx_15_2_short():
     # Test which channels are short
     # These are the ones marked as red at
     # https://github.com/mne-tools/mne-testing-data/pull/51 step 4 figure 2
-    is_short = nirs_short_channels(raw.info)
+    is_short = short_channels(raw.info)
     assert_array_equal(is_short[:9:2], [False, True, False, True, False])
-    is_short = nirs_short_channels(raw.info, threshold=0.003)
+    is_short = short_channels(raw.info, threshold=0.003)
     assert_array_equal(is_short[:3:2], [False, False])
-    is_short = nirs_short_channels(raw.info, threshold=50)
+    is_short = short_channels(raw.info, threshold=50)
     assert_array_equal(is_short[:3:2], [True, True])
 
     # Test trigger events
@@ -183,7 +183,7 @@ def test_nirx_15_0():
 
     # Test distance between optodes matches values from
     allowed_distance_error = 0.0002
-    distances = nirs_source_detector_distances(raw.info)
+    distances = source_detector_distances(raw.info)
     assert_allclose(distances[::2], [
         0.0301, 0.0315, 0.0343, 0.0368, 0.0408,
         0.0399, 0.0393, 0.0367, 0.0336, 0.0447], atol=allowed_distance_error)
