+++ mne/preprocessing/nirs/__init__.py
@@ -1,3 +1,6 @@
 """NIRS specific preprocessing functions."""
 
-from .nirs import nirs_short_channels, nirs_source_detector_distances
+from .nirs import short_channels, source_detector_distances
+from ._optical_density import optical_density
+from ._beer_lambert_law import beer_lambert_law
+
